Revert "Reapply "Cleanup2 (#3501)" (#3503) (#3504)"
This reverts commit 798f67f6f52a5ca01ce92971c8498bfcda565709.

+++ src/test/java/ti4/model/StrategyCardSetModelTest.java
@@ -33,6 +33,6 @@ public class StrategyCardSetModelTest extends BaseTi4Test {
                         .map(StrategyCardModel::getInitiative)
                         .distinct()
                         .count()
-                == scSetModel.getScIDs().size();
+                == (long) scSetModel.getScIDs().size();
     }
 }
